+++ hqet_heavy_to_light/lattice.h
@@ -119,7 +119,7 @@ enum quark_type_choices { STANDARD_QUARK , STATIC_LIGHT_QUARK  } ;
 
 /* The following are global scalars */
 EXTERN	int nx,ny,nz,nt;	/* lattice dimensions */
-EXTERN  int volume;	        /* volume of lattice = nx*ny*nz*nt */
+EXTERN  size_t volume;	        /* volume of lattice = nx*ny*nz*nt */
 
 EXTERN	Real clov_c,u0;
 EXTERN  Real byterevReal;  /* flag to bytereverse lattice */
@@ -199,7 +199,7 @@ enum layout_flag_set { HYPER_GRAY_EVENFIRST = 100 } ;
 
 /* Some of these global variables are node dependent */
 /* They are set in "make_lattice()" */
-EXTERN  int	sites_on_node;		/* number of sites on this node */
+EXTERN  size_t	sites_on_node;		/* number of sites on this node */
 EXTERN  int	even_sites_on_node;	/* number of even sites on this node */
 EXTERN  int	odd_sites_on_node;	/* number of odd sites on this node */
 EXTERN  int	number_of_nodes;	/* number of nodes in use */
